refactor: HeaderArea에서 Header, Banner 분리 및 스타일링
