--- src/components/HeaderArea/Banner.js
@@ -1,14 +1,0 @@
-const Banner = () => {
-  return /* html */ `
-    <div class="top-rated-movie">
-      <div class="rate">
-        <img src="./images/star_empty.png" class="star" />
-        <span class="rate-value">9.5</span>
-      </div>
-      <div class="title">인사이드 아웃2</div>
-      <button class="primary detail">자세히 보기</button>
-    </div>
-  `;
-};
-
-export default Banner;
