+++ src/styles/main.css
@@ -44,6 +44,7 @@ button.primary {
 }
 
 #wrap {
+  position: relative;
   min-width: 1440px;
   background-color: var(--color-bluegray-100);
 }
@@ -75,7 +76,15 @@ button.primary {
   background-color: rgba(0, 0, 0, 0.5);
   width: 100%;
   height: 100%;
-  z-index: 1;
+}
+
+.header {
+  position: fixed;
+  top: 40px;
+  left: calc(50% - 640px);
+  z-index: 9999;
+  width: 100%;
+  max-width: 1280px;
 }
 
 .header-container {
@@ -123,6 +132,7 @@ button.primary {
   z-index: 2;
   max-width: 1280px;
   margin: 0 auto;
+  padding-top: 74.5px;
 }
 
 .top-rated-movie {
